requirements.txt
requests
Django==2.2.4
django-crispy-forms==1.7.2
django-jsonify==0.3.0
django-static-fontawesome==5.0.8.1
django-taggit==1.1.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
